refactor: ♻️ 환경 표시 상수 공용화 및 세션 테스트 분리
