+++ apps/admin/src/lib/auth/session.test.ts
@@ -1,0 +1,44 @@
+import { beforeEach, describe, expect, it, vi } from "vitest";
+
+const { removeAccessToken, removeAdminApiEnvironment, saveAdminApiEnvironment } = vi.hoisted(() => ({
+	removeAccessToken: vi.fn(),
+	removeAdminApiEnvironment: vi.fn(),
+	saveAdminApiEnvironment: vi.fn(),
+}));
+
+vi.mock("@/lib/api/auth", () => ({
+	reissueAccessTokenApi: vi.fn(),
+}));
+
+vi.mock("@/lib/utils/localStorage", () => ({
+	loadAccessToken: vi.fn(),
+	removeAccessToken,
+	removeAdminApiEnvironment,
+	saveAccessToken: vi.fn(),
+	saveAdminApiEnvironment,
+}));
+
+describe("switchAdminApiEnvironment", () => {
+	beforeEach(() => {
+		removeAccessToken.mockReset();
+		removeAdminApiEnvironment.mockReset();
+		saveAdminApiEnvironment.mockReset();
+	});
+
+	it("세션(access token, 환경 값)을 모두 지운 뒤 새 환경을 저장한다", async () => {
+		const { switchAdminApiEnvironment } = await import("./session");
+		const redirect = vi.fn();
+		const callOrder: string[] = [];
+
+		removeAdminApiEnvironment.mockImplementation(() => callOrder.push("removeAdminApiEnvironment"));
+		saveAdminApiEnvironment.mockImplementation(() => callOrder.push("saveAdminApiEnvironment"));
+
+		switchAdminApiEnvironment("prod", redirect);
+
+		expect(removeAccessToken).toHaveBeenCalledOnce();
+		expect(removeAdminApiEnvironment).toHaveBeenCalledOnce();
+		expect(saveAdminApiEnvironment).toHaveBeenCalledWith("prod");
+		expect(callOrder).toEqual(["removeAdminApiEnvironment", "saveAdminApiEnvironment"]);
+		expect(redirect).toHaveBeenCalledWith("/auth/login");
+	});
+});
